requirements.txt
python-dotenv
litellm
PyPDF2